requirements.txt
bcrypt==4.0.1
boto3==1.34.0
botocore==1.34.162
fastapi==0.104.1
httpx==0.28.1
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
uvicorn[standard]==0.24.0